+++ src/commands/fun.ts
@@ -1,51 +1,42 @@
-import { Client, Discord, Guard, SlashOption, Slash } from 'discordx';
-import { CommandInteraction, GuildMember, User } from 'discord.js';
+import { Client, Discord, Guard, SlashOption, Slash, ContextMenu } from 'discordx';
+import { CommandInteraction, GuildMember, User, UserContextMenuInteraction } from 'discord.js';
 
 import IsConfigEnabled from '../guards/config/is-config-enabled';
 import ServerExists from '../guards/config/server-exists';
-import { getOrCreateCitizen } from '../db/get-or-create-citizen';
 import GuardCache from '../types/GuardCache';
+import { ding, unding } from '../helpers/ding-helpers';
+import { gulag, ungulag } from '../helpers/gulag-helpers';
 
 @Discord()
 export default class Fun {
+  @ContextMenu('USER', 'ding')
+  async dingUser(interaction: UserContextMenuInteraction) {
+    await ding(interaction.targetUser, interaction);
+  }
+
+  @ContextMenu('USER', 'unding')
+  async undingUser(interaction: UserContextMenuInteraction) {
+    await unding(interaction.targetUser, interaction);
+  }
+
   @Slash('ding', {
     description: 'Make a game out of shaming someone for talking with their mic muted',
   })
-  async ding(
+  async dingCommand(
     @SlashOption('user', { type: 'USER' })
     user: User,
     interaction: CommandInteraction
   ): Promise<void> {
-    const member = await interaction.guild.members.fetch(user);
-    if (!member) {
-      await interaction.reply(`Could not find that member`);
-      return;
-    }
-    const citizen = await getOrCreateCitizen(member.id);
-    citizen.dings++;
-    await citizen.save();
-    await interaction.reply(
-      `DING!\n${member} has now talked ${citizen.dings} times with their mic muted`
-    );
+    await ding(user, interaction);
   }
 
   @Slash('unding', { description: 'When you ding someone by mistake' })
-  async unding(
+  async undingCommand(
     @SlashOption('user', { type: 'USER' })
     user: User,
     interaction: CommandInteraction
   ): Promise<void> {
-    const member = await interaction.guild.members.fetch(user);
-    if (!member) {
-      await interaction.reply(`Could not find that member`);
-      return;
-    }
-    const citizen = await getOrCreateCitizen(member.id);
-    citizen.dings--;
-    await citizen.save();
-    await interaction.reply(
-      `Someone did a whoopsie! Ding count for ${member} is now ${citizen.dings}.`
-    );
+    await unding(user, interaction);
   }
 
   @Slash('rr', { description: `Do it, you won't, no balls` })
@@ -72,18 +63,17 @@ export default class Fun {
     }
 
     const hit = roll > 5 / 6;
-    const hadMemberRole = interaction.member.roles.cache.has(memberRole);
-    const hadGulagRole = interaction.member.roles.cache.has(gulagRole);
+
+    const isInGulag = interaction.member.roles.cache.has(gulagRole);
+
+    if (isInGulag) {
+      await interaction.reply(`Sorry kiddo, there's no playing games in the gulag.`);
+      return;
+    }
 
     let res: string;
     if (hit) {
-      if (hadMemberRole) {
-        promises.push(interaction.member.roles.remove(memberRole));
-      }
-      promises.push(
-        interaction.member.roles.add(gulagRole),
-        interaction.member.voice.setMute(true)
-      );
+      promises.push(gulag(interaction.user, server, interaction));
       res = `**BANG!** You're dead, ${interaction.member}.`;
     } else {
       res = `**CLICK!** You lived, ${interaction.member}! For now...`;
@@ -95,13 +85,7 @@ export default class Fun {
     // remove role after 10 seconds
     if (hit && interaction.member) {
       await new Promise((resolve) => setTimeout(resolve, mutedSeconds * 1000));
-      if (!hadGulagRole) {
-        await interaction.member.roles.remove(gulagRole);
-      }
-      if (hadMemberRole) {
-        await interaction.member.roles.add(memberRole);
-      }
-      await interaction.member.voice.setMute(false);
+      await ungulag(interaction.user, server, interaction);
     }
   }
 }