+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.2.3
  - add context menu items for gulag (right-click -> apps -> gulag)
+ - add context menu items for dings
 
 ## 1.2.2
  - fix types for command arguments